test_instance_seg_teeth/requirements.txt
# 牙齿实例分割模块依赖

# 深度学习框架
tensorflow>=2.10.0
keras>=2.10.0

# 目标检测
ultralytics>=8.0.28

# 图像处理
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-image>=0.19.0

# 数值计算
numpy>=1.21.0

# 可选：用于数据可视化
matplotlib>=3.5.0

# 可选：用于性能监控
psutil>=5.8.0
